requirements.txt
Flask==3.0.0
opencv-python==4.9.0.80
numpy>=1.26.0
Pillow==10.1.0
scikit-learn>=1.3.0
matplotlib==3.8.2
Werkzeug==3.0.1